+++ lib/output/verbose_output.py
@@ -22,8 +22,7 @@ import time
 from threading import Lock
 from urllib.parse import urlparse
 
-from lib.utils.data import human_size
-from lib.utils.terminal_size import get_terminal_size
+from lib.utils.size import human_size, get_terminal_size
 from .colors import ColorOutput
 
 if sys.platform in ["win32", "msys"]:
